+++ lib/braintree/test/transaction_amounts.rb
@@ -3,6 +3,7 @@ module Braintree
     # NEXT_MAJOR_VERSION are these even being used anymore? Can we remove this class??
     module TransactionAmounts
       Authorize = "1000.00"
+      PartiallyAuthorized = "1004.00"
       Decline = "2000.00"
       HardDecline = "2015.00"
       Fail = "3000.00"